Rename scripts and mark them as such
Mark actual top level scripts as such instead of disguising them as modules.

Also remove __init__.py from ipsilon/install as this is not a module just
the place where install scripts are kept, for now.

Note: Scripts are installed in the bin directory but the contrib spec file
moves them to sbin.

Signed-off-by: Simo Sorce <simo@redhat.com>