Change provider plugins registration and enablement
When plugins are not enabled at startup the admin page is not available
as it is created only on enablement.

Split enablement and registration, so plugins can be registered even
when actually disabled.

Also rework the way enablement is tracked and make sure enablement status
is saved back to the database when it changes so it is kept on restarts.

Signed-off-by: Simo Sorce <simo@redhat.com>