Properly identify code location of logging message
The method Log.call_location() is used to add identifying infomation
about the location in the code where a logging message is emitted
from.  It needs to walk up the stack to bypass calls involved in
logging to find where the call to logging was made. Formerly the code
has a hardcoded offset into the list of stack frame objects. But any
change in the logging implementation perturbs that offset. This patch
fixes that problem by walking up the stack until a non-logging
function is identified.

Ticket: 172
Signed-off-by: John Dennis <jdennis@redhat.com>
Reviewed-by: Patrick Uiterwijk <puiterwijk@redhat.com>