Transaction object must always have a transaction id
A Transaction object must always have a transaction id (tid) because a
tid is how a transaction is referenced (e.g. when being stored or
retrieved). The existing code erroneously assumed a tid only needed to
be created if the request contained query or form parameters. This
restriction is now removed, if the transaction id cannot be found as a
request parameter a tid is unconditionally created.

Ticket: 177
Signed-off-by: John Dennis <jdennis@redhat.com>
Reviewed-by: Patrick Uiterwijk <puiterwijk@redhat.com>