Verify that uploaded files are readable when selected
When selecting a file using file input this doesn't guarantee
that the file contents are readable. No real error is provided
in this case, things just don't work as expected.

When adding a SAML SP and the metadata file or image is provided via
file upload try to read the file in the browser first and pop up an
alert if it is unreadable.

https://fedorahosted.org/ipsilon/ticket/22

Signed-off-by: Rob Crittenden <rcritten@redhat.com>
Reviewed-by: Simo Sorce <simo@redhat.com>
Reviewed-by: John Dennis <jdennis@redhat.com>