ipsilon-server-install sometimes fails to log & emit errors
ipsilon-server-install may silently and immediately fail, nothing is
emitted to the console nor captured in the log file, it's just a
silent complete failure.

An example that reproduces the problem is a hostname without any dots
in it, e.g. "localhost".

The log level is set after some code executes (e.g. arg parsing).
If that code raises an error the exception handler will log it
at the debug level, but because the log level has not been set
yet to debug (it's still at the default error level) the message
is not emitted.

The log level should be set as soon as logging is initialized.

An error message should be emitted to the console, therefore in
additon to the exception handler logging the error to the debug log
along with the stack trace it should also emit just the message to the
console.

Ticket: 202
Signed-off-by: John Dennis <jdennis@redhat.com>
Reviewed-by: Patrick Uiterwijk <puiterwijk@redhat.com>