Allow SAML SP metadata to be viewed and edited
After an SP is created it is impossible to view the metadata
associated with the SP. Being able to view the metadata is
essential for diagnosing problems because the SP metadata
is the configuration for the SP. It is also useful to be
able to update the SP's metadata.

Add Metadata as a config item, for some reason it never existed. The
Metadata config item must be present for the options_config.html
template can access it.

Assure only the admin can update the metadata.

Add a new utility function validate_sp_metadata() that attempts
to assure the metadata is valid, call this from all the locations
where metadata is loaded.

Add the 'multiline' property to the String config class, multiline
directs options_config.html to generate a textarea widget instead of
single line text input.

Ticket: 130
Signed-off-by: John Dennis <jdennis@redhat.com>
Reviewed-by: Rob Crittenden <rcritten@redhat.com>