Initialize plugins outside of the upgrade cycle
This will make sure that cleanups actually get scheduled for plugin
datastores.

Callers were always expected to catch any errors in the initialization of plugins.
The try/catch that create_plugin_data_table() is moved out of never was catching the
errors for any other reason than to make sure we don't leave the function call
with _upgrade set to True.

Ticket: #209
Fixes: #76
Signed-off-by: Patrick Uiterwijk <puiterwijk@redhat.com>
Reviewed-by: Rob Crittenden <rcritten@redhat.com>