Remove readonly check when only getting plugin information
This fixes the homepage when using file-based configuration.
The homepage was broken because ipsilon/root.py used _get_plugin_data
to get a reference to the currently enabled provider plugins, but this
function threw an error in case the configuration was readonly.
This check was too hard, since if you are only getting information,
this call is perfectly reasonable to use.

Ticket: #210
Fixes: #77
Signed-off-by: Patrick Uiterwijk <puiterwijk@redhat.com>
Reviewed-by: Rob Crittenden <rcritten@redhat.com>