Error out if requested to send newline via OpenID AX
The kv serialization format of OpenID used for verification in stateless
protocol runs, attributes values containing newlines are impossible.
Previously, we would throw a 500 because the python-openid library throws
an uncaught exception, but now we error out at the moment the user is
still at us, to allow for them to inform admins.

Signed-off-by: Patrick Uiterwijk <puiterwijk@redhat.com>
Reviewed-by: Pierre-Yves Chibon <pingou@pingoured.fr>