Move the KDC port outside of the privileged range
On EL6 we can't use the libsocket-wrapper. This means that
ports below 1024 are still privileged, and this would mean
we would need to run the test suite as root.
Since we explicitly say we do not want to run this as root,
let's instead just move the last thing to a non-privileged
port.

Signed-off-by: Howard Johnson <merlin@merlinthp.org>
Signed-off-by: Patrick Uiterwijk <puiterwijk@redhat.com>
Reviewed-by: Pierre-Yves Chibon <pingou@pingoured.fr>